small tweaks from old master

+++ src/DP_Player.py
@@ -831,7 +831,7 @@ class DP_Player(InfoBarBase, InfoBarShowHide, InfoBarCueSheetSupport,
 		totalTime = self.getPlayLength()[1] / 90000
 
 		if currentTime is not None and currentTime > 0 and totalTime is not None and totalTime > 0:
-			progress = currentTime / (totalTime/100)
+			progress = currentTime / float(totalTime/100.0)
 			printl( "played time is %s secs of %s @ %s%%" % ( currentTime, totalTime, progress),self, "I" )
 		else:
 			progress = 0
